TwoStageOptimalControl.integ4 — Function

integ4(n::Int64, h::Float64, F)

integ4 performs 4th order integration (Simpson-method) method for datavector F with n points with a distance of h between each of them. If n > 1 and F = 1, F will be interpreted as a constant function over the n points.

TwoStageOptimalControl.integ2 — Function

integ2(n::Int64, h::Float64, F)

integ2 performs 2nd order integration (trapezoid) method for datavector F with n points with a distance of h between each of them. If n > 1 and F = 1, F will be interpreted as a constant function over the n points.

TwoStageOptimalControl.integ1 — Function

integ1(n::Int64, h::Float64, F)

integ1 performs 1st order left boundary integration method for datavector F with n points with a distance of h between each of them. If n > 1 and F = 1, F will be interpreted as a constant function over the n points.

TwoStageOptimalControl.CumInteg4 — Function

CumInteg4(n::Int64, h::Float64, F)

CumInteg4 calculates the cumulative integrals

\[\int_0^t F(s)ds \forall t\in(0,T)\]

of F using a 4th order integration method.

n -> number of datapoints,

h -> distance between datapoints.

TwoStageOptimalControl.interp1 — Function

interp1(xpt, ypt; method = "Linear", extrapolation = Flat())

One dimensional interpolation using either linear interpolation or BSplines

TwoStageOptimalControl.interp2 — Function

interp2(xpt,ypt,zpt)

Two dimensional interpolation using BSplines

TwoStageOptimalControl.SavitskyGolaySmoothing — Function

SavitskyGolaySmoothing(x::Vector, windowSize::Integer, polyOrder::Integer; deriv::Integer=0)

Smooth a vector x using the Savitsky-Golay method

TwoStageOptimalControl.ParaAdjust — Function

ParaAdjust(h_global::Float64, Para::Dict)

ParaAdjust adjust all relevant elements of the Parameter dictionary Para to the step-size h_global.

TwoStageOptimalControl.LoadVariables — Function

LoadVariables(Para, Results)

Load inital profiles for the control variables from the Results dictionary and use interpolation if necessary

TwoStageOptimalControl.ConInterpol — Function

ConInterpol(Con_dist::Array{Float64,3}, Con::Array{Float64,3}, Para::Dict)

Given Controls Con_dist and Con, ConInterpol performs an interpolation of these fields to fit the passed Parameter dict Para.

TwoStageOptimalControl.ConSmooth — Function

ConSmooth(Con::Array{Float64,3}, Con_dist::Array{Float64,3}, Para::Dict)

Smooth the profiles of the control variables using a SavitskyGolaySmoother

TwoStageOptimalControl.GradSmooth — Function

GradSmooth(dHam::Array{Float64,3}, dHam_dist::Array{Float64,3}, Para::Dict)

Smooth the gradient by using the median value of a moving window.